requirements_hybrid.txt
# FastAPI Hybrid YouTube Summarizer Dependencies

# Web Framework
fastapi==0.112.0
uvicorn[standard]==0.30.5

# AI/ML
google-generativeai==0.7.2

# HTTP & API
requests==2.31.0
httpx==0.27.0

# Environment & Configuration
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0

# Logging & Utilities
python-multipart==0.0.6

# Cloud Run compatibility
gunicorn==21.2.0